app/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
sqlalchemy==2.0.23
pymysql==1.1.0
cryptography==41.0.5
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
pydantic==2.0.3
pydantic-settings==2.0.3
python-dotenv==1.0.0
fastapi-mail==1.4.1
openai>=1.18.0
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.0
joblib==1.4.2
apscheduler==3.10.4